Reorganise packages
Reorganise the packages to:

- make the names more idiomatic in Go

- move internal packages to the `internal` directory to prevent someone
  from importing them (essentially making them 'private' packages
  outside of this code base)

I used these references when deciding what structure and naming to use:

https://blog.golang.org/package-names
https://rakyll.org/style-packages/
https://peter.bourgon.org/go-best-practices-2016/
https://medium.com/@benbjohnson/standard-package-layout-7cdbc8391fc1
https://docs.google.com/document/d/1e8kOo3r51b2BWtTs_1uADIA5djfXhPT36s6eHVRIvaU/edit

I still need to:

- move most of main.go into `cmd/athensdb/main.go`

- consider moving `acceptance_test.go`, which is in the `main` package,
  but maybe doesn't belong in `cmd/athensdb`
