+++ internal/integration/Dockerfile
@@ -4,7 +4,7 @@ MAINTAINER Matt Bostock <matt@mattbostock.com>
 WORKDIR /go/src/github.com/mattbostock/athensdb
 COPY . /go/src/github.com/mattbostock/athensdb
 
-RUN go test -c -o /bin/athensdb.test -race -tags integration ./integration_tests && \
+RUN go test -c -o /bin/athensdb.test -race -tags integration ./internal/integration && \
   cd /bin && \
   rm -rf /go/src
