+++ main.go
@@ -17,8 +17,8 @@ import (
 	"github.com/golang/protobuf/proto"
 	"github.com/golang/snappy"
 	"github.com/hashicorp/memberlist"
-	v1API "github.com/mattbostock/athensdb/api/v1"
-	"github.com/mattbostock/athensdb/remote"
+	v1API "github.com/mattbostock/athensdb/internal/api/v1"
+	"github.com/mattbostock/athensdb/internal/remote"
 	"github.com/prometheus/common/log"
 	"github.com/prometheus/common/route"
 	"github.com/prometheus/prometheus/pkg/labels"
